Move engine code to core directory
