+++ core/entity.py
