+++ core/entity_manager.py
