+++ core/scene.py
